requirements.txt
ipynb>=0.5.1
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
git+https://github.com/facebookresearch/segment-anything.git
scikit-learn>=1.2.0
tqdm>=4.65.0